Split the tests up by producer
This should help with maintainability, as `MessageQueuesSpec` was
starting to get long and unweildy.
